+++ src/flutter_ffi.rs
@@ -94,8 +94,8 @@ pub fn stop_global_event_stream(app_type: String) {
 }
 pub enum EventToUI {
     Event(String),
-    Rgba(usize),
-    Texture(usize, bool), // (display, gpu_texture)
+    Rgba(i32),
+    Texture(i32, bool), // (display, gpu_texture)
 }
 
 pub fn host_stop_system_key_propagate(_stopped: bool) {
